requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
Click==7.0
colorama==0.4.1
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29.3
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
Flask==1.0.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.0
opencv-contrib-python==4.0.0.21
pandocfilters==1.4.2
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-hub==0.2.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wincertstore==0.2
